bpf: Add proxy delegation flag
Add a new proxy delegation flag called SVC_FLAG_L7_DELEGATE. This
flag aliases to SVC_FLAG_LOOPBACK but does not overlap in its use.
The latter is only used in combination with HostPort, and the control
plane makes sure that there is no collision.

Reusing SVC_FLAG_L7_LOADBALANCER was not an option since this collides
with session affinity / per-service algorithm. For services marked
with SVC_FLAG_L7_DELEGATE, the behavior is that in lb{4,6}_local() we
check whether the selected backend is local on the node, and if it is,
then the packet is punted up to a user space proxy for processing. If
the backend is non-local, then the dataplane forwards the traffic as
per underlying KPR configuration to the remote backend.

This is different to SVC_FLAG_L7_LOADBALANCER since in that case all
traffic no matter the backend is punted up to a proxy. The new flag
SVC_FLAG_L7_DELEGATE therefore can help scalability e.g. going from
L3/4 up to L7 since the former can be handled in tcx/XDP.

For consistency, also do a few renames SVC_FLAG_L7LOADBALANCER into
SVC_FLAG_L7_LOADBALANCER, and *lb{4,6}_svc_is_l7loadbalancer() into
*lb{4,6}_svc_is_l7_loadbalancer().

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>